requirements.txt
h5py==3.3.0
libpysal==4.5.1
matplotlib==3.3.2
numpy==1.19.2
openslide_python==1.1.2
pandas==1.2.1
Pillow==10.4.0
scikit_learn==0.23.2
scipy==1.1.0
scipy==1.5.2
seaborn==0.9.0
seaborn==0.11.2
torch==1.7.1
torch_geometric==1.7.2
torchvision==0.8.2
